Tareas/Tarea2/app/requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
filetype==1.2.0
Flask==2.3.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pillow==10.3.0
pycparser==2.21
PyMySQL==1.1.0
Werkzeug==2.3.7
zipp==3.17.0
